Add shortcut for lazy loading chunks
