+++ src/api/Settings.ts
@@ -129,7 +129,7 @@ export const SettingsStore = new SettingsStoreClass(settings, {
 
         if (path === "plugins" && key in plugins)
             return target[key] = {
-                enabled: plugins[key].required ?? plugins[key].enabledByDefault ?? false
+                enabled: IS_REPORTER ?? plugins[key].required ?? plugins[key].enabledByDefault ?? false
             };
 
         // Since the property is not set, check if this is a plugin's setting and if so, try to resolve
