+++ src/plugins/partyMode/index.ts
@@ -18,7 +18,7 @@
 
 import { definePluginSettings, migratePluginSettings } from "@api/Settings";
 import { Devs } from "@utils/constants";
-import definePlugin, { OptionType } from "@utils/types";
+import definePlugin, { OptionType, ReporterTestable } from "@utils/types";
 import { FluxDispatcher } from "@webpack/common";
 
 const enum Intensity {
@@ -46,6 +46,7 @@ export default definePlugin({
     name: "PartyMode",
     description: "Allows you to use party mode cause the party never ends ✨",
     authors: [Devs.UwUDev],
+    reporterTestable: ReporterTestable.None,
     settings,
 
     start() {
